Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,8 +1,8 @@
 #include<stdio.h>
 int main(){
-int x,y,z;
+float x,y,z;
 float a;
-scanf("%d %d %d ", &x , &y , &z);
+scanf("%f %f %f ", &x , &y , &z);
 a = (x+y+z)/3;
 printf("Average: %.2f",a);
 }
